R removed some default comments at begging of new files R separated content of redisStorage in multiple files and placed under its folder
